requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
contourpy==1.3.3
cycler==0.12.1
fastapi==0.116.1
fonttools==4.59.1
gitdb==4.0.12
GitPython==3.1.45
greenlet==3.2.4
h11==0.16.0
httptools==0.6.4
idna==3.10
importlib-metadata==6.11.0
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
kiwisolver==1.4.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.8.1
mdurl==0.1.2
narwhals==2.1.2
numpy==1.26.4
nvidia-nccl-cu12==2.27.7
packaging==23.2
pandas==2.1.3
pillow==10.4.0
plotly==6.3.0
protobuf==4.25.8
psycopg2-binary==2.9.7
pyarrow==21.0.0
pydantic==2.2.1
pydantic_core==2.6.1
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.6
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.5
rich==13.9.4
rpds-py==0.27.0
scikit-learn==1.3.2
scipy==1.16.1
seaborn==0.13.2
setuptools==78.1.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.20
starlette==0.47.2
streamlit==1.28.0
tenacity==8.5.0
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.2
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.23.2
uvloop==0.21.0
validators==0.35.0
watchdog==6.0.0
watchfiles==1.1.0
websockets==15.0.1
wheel==0.45.1
xgboost==2.1.1
xlrd @ file:///croot/xlrd_1685030938141/work
zipp==3.23.0
